refactor(client): P5b(4/5) — media/file handlers to native_tools/media.py
Verbatim moves: browser_screenshot, generate_file, post_file,
analyze_image (vision marker dict preserved), generate_image,
_detect_image_type (bot keeps a staticmethod alias). get_config
provider-backed; asyncio.TimeoutError normalized to TimeoutError in the
moved post_file body (builtin alias, new-file lint).
